requirements.txt
asgiref==3.4.1   
Django==3.2.6   
django-cors-headers==3.8.0   
djangorestframework==3.12.4
mysqlclient==2.0.3
Pillow==8.3.1
pip==21.2.4
pytz==2021.1
setuptools==41.2.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
react-beforeunload